+++ Vungle/src/main/java/com/mopub/mobileads/VungleMediationConfiguration.java
@@ -80,9 +80,11 @@ public class VungleMediationConfiguration implements MediationSettings {
         if (extras.containsKey(Builder.EXTRA_START_MUTED_KEY)) {
             final String isStartMuted = extras.get(Builder.EXTRA_START_MUTED_KEY);
             adConfig.setMuted(Boolean.parseBoolean(isStartMuted));
-        } else {
+        } else if (extras.containsKey(Builder.EXTRA_SOUND_ENABLED_KEY)) {
             final String isSoundEnabled = extras.get(Builder.EXTRA_SOUND_ENABLED_KEY);
-            adConfig.setMuted(!Boolean.parseBoolean(isSoundEnabled));
+            if (!TextUtils.isEmpty(isSoundEnabled)) {
+                adConfig.setMuted(!Boolean.parseBoolean(isSoundEnabled));
+            }
         }
 
         final String flexViewCloseTimeInSec = extras.get(Builder.EXTRA_FLEXVIEW_CLOSE_TIME_KEY);