+++ Vungle/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+  * 6.7.0.0
+    * This version of the adapters has been certified with Vungle 6.7.0 and MoPub SDK 5.13.0.
+    * Add support for Advanced Bidding for fullscreen ads.
+    * Add `onAdClick` callback support.
+
   * 6.5.3.1
     * Refactor non-native adapter classes to use the new consolidated API from MoPub.
     * This and newer adapter versions are only compatible with 5.13.0+ MoPub SDK.